Store the price the customer was shown, not one re-derived from it
spQuoteLine_Insert derived NetPrice from ListPrice and an integer
DiscountPct unconditionally. That was right for the only caller it had —
an admin typing numbers into the portal — and wrong for anything
carrying a price that came out of PriceResolver.

The reachable half is rounding. The old expression did none, and money
carries four decimal places, so a list price of 99.99 at 7% stored
92.9907 while the page displayed 92.99. That is a price nobody can pay,
on the document the customer signs, and it needs no unusual
configuration to happen.

The other half needs a store to set MinMarginPct. A price held up by the
floor is cost x (1 + margin), and the discount that produces it is
whatever it happens to be — PriceResolver returns EffectiveDiscountPct
as a decimal for that reason, while QuoteLine.DiscountPct was an INT. A
floored line recorded there would print a discount that does not produce
the price beside it. (A group's own discount is integral:
CustomerGroup.Discount is an INT and fnCatalog_VisibleProducts takes
@DiscountPct int. The floor is the only source of a fractional effective
discount today.)

So the caller states the net price and the procedure derives one only
when it did not. The admin path is untouched; the storefront path will
be explicit. Where it does derive, it now uses PriceResolver's exact
shape, which is also fnCatalog_VisibleProducts': cast out of money
before multiplying, then ROUND to two places — T-SQL's ROUND is
half-away-from-zero, which is MidpointRounding.AwayFromZero and not the
framework default, so two lines of one product cannot round differently.

This is the existing parity rule carried one step further. CLAUDE.md
already said SQL computes an ordering key and PriceResolver computes the
price a customer is shown; the addition is that the price a customer was
shown is the price that gets stored. QuoteLinePriceTests drives 56
combinations of list price and discount through the derived path and
asserts it agrees with PriceResolver to the cent, the way
CatalogPriceParityTests does for the sort key.

QuoteLine.DiscountPct widens to DECIMAL(5, 2), with QuoteLineModel, the
API contract and the portal's display following. Integers widen
losslessly, so no existing row or caller moves.

Verified against the development database: 81/81 in the library suite,
100 in the API suite, 107 in SMStore, 58 in SMPortal.

+++ SMDataManager.Library/Models/QuoteLineModel.cs
@@ -9,7 +9,12 @@ namespace SMDataManager.Library.Models
         public string Name { get; set; }
         public int Quantity { get; set; }
         public decimal ListPrice { get; set; }
-        public int DiscountPct { get; set; }
+        /// <summary>The discount that produced <see cref="NetPrice"/>, to two places.</summary>
+        /// <remarks>
+        /// Decimal because a price held up by the site's margin floor has a fractional
+        /// effective discount — see spQuoteLine_Insert.
+        /// </remarks>
+        public decimal DiscountPct { get; set; }
         public decimal NetPrice { get; set; }
     }
 }
